requirements.txt
Django==3.0.3
djangorestframework==3.11.0
flake8==3.7.9
psycopg2==2.8.4
sentry-sdk==0.14.2